+++ lib/framework/AuthenticationService.cpp
@@ -6,7 +6,7 @@
  *   https://github.com/theelims/ESP32-sveltekit
  *
  *   Copyright (C) 2018 - 2023 rjwats
- *   Copyright (C) 2023 theelims
+ *   Copyright (C) 2023 - 2024 theelims
  *
  *   All Rights Reserved. This software may be modified and distributed under
  *   the terms of the LGPL v3 license. See the LICENSE file for details.
@@ -27,13 +27,13 @@ void AuthenticationService::begin()
     _server->on(SIGN_IN_PATH, HTTP_POST, [this](PsychicRequest *request, JsonVariant &json)
                 {
         if (json.is<JsonObject>()) {
-            String         username       = json["username"];
-            String         password       = json["password"];
+            String username = json["username"];
+            String password = json["password"];
             Authentication authentication = _securityManager->authenticate(username, password);
             if (authentication.authenticated) {
-                PsychicJsonResponse response = PsychicJsonResponse(request, false, 256);
-                JsonObject          root     = response.getRoot();
-                root["access_token"]         = _securityManager->generateJWT(authentication.user);
+                PsychicJsonResponse response = PsychicJsonResponse(request, false);
+                JsonObject root = response.getRoot();
+                root["access_token"] = _securityManager->generateJWT(authentication.user);
                 return response.send();
             }
         }