+++ lib/framework/BatteryService.cpp
@@ -5,7 +5,7 @@
  *   with responsive Sveltekit front-end built with TailwindCSS and DaisyUI.
  *   https://github.com/theelims/ESP32-sveltekit
  *
- *   Copyright (C) 2023 theelims
+ *   Copyright (C) 2023 - 2024 theelims
  *
  *   All Rights Reserved. This software may be modified and distributed under
  *   the terms of the LGPL v3 license. See the LICENSE file for details.
@@ -24,7 +24,7 @@ void BatteryService::begin()
 
 void BatteryService::batteryEvent()
 {
-    StaticJsonDocument<32> doc;
+    JsonDocument doc;
     char message[32];
     doc["soc"] = _lastSOC;
     doc["charging"] = _isCharging;