+++ lib/framework/EventEndpoint.h
@@ -9,7 +9,7 @@
  *   https://github.com/theelims/ESP32-sveltekit
  *
  *   Copyright (C) 2018 - 2023 rjwats
- *   Copyright (C) 2023 theelims
+ *   Copyright (C) 2023 - 2024 theelims
  *
  *   All Rights Reserved. This software may be modified and distributed under
  *   the terms of the LGPL v3 license. See the LICENSE file for details.
@@ -27,13 +27,11 @@ public:
     EventEndpoint(JsonStateReader<T> stateReader,
                   JsonStateUpdater<T> stateUpdater,
                   StatefulService<T> *statefulService,
-                  EventSocket *socket, const char *event,
-                  size_t bufferSize = DEFAULT_BUFFER_SIZE) : _stateReader(stateReader),
-                                                             _stateUpdater(stateUpdater),
-                                                             _statefulService(statefulService),
-                                                             _socket(socket),
-                                                             _bufferSize(bufferSize),
-                                                             _event(event)
+                  EventSocket *socket, const char *event) : _stateReader(stateReader),
+                                                            _stateUpdater(stateUpdater),
+                                                            _statefulService(statefulService),
+                                                            _socket(socket),
+                                                            _event(event)
     {
         _statefulService->addUpdateHandler([&](const String &originId)
                                            { syncState(originId); },
@@ -53,7 +51,6 @@ private:
     StatefulService<T> *_statefulService;
     EventSocket *_socket;
     const char *_event;
-    size_t _bufferSize;
 
     void updateState(JsonObject &root, int originId)
     {
@@ -62,7 +59,7 @@ private:
 
     void syncState(const String &originId, bool sync = false)
     {
-        DynamicJsonDocument jsonDocument{_bufferSize};
+        JsonDocument jsonDocument;
         JsonObject root = jsonDocument.to<JsonObject>();
         String output;
         _statefulService->read(root, _stateReader);