+++ lib/framework/FeaturesService.cpp
@@ -6,7 +6,7 @@
  *   https://github.com/theelims/ESP32-sveltekit
  *
  *   Copyright (C) 2018 - 2023 rjwats
- *   Copyright (C) 2023 theelims
+ *   Copyright (C) 2023 - 2024 theelims
  *
  *   All Rights Reserved. This software may be modified and distributed under
  *   the terms of the LGPL v3 license. See the LICENSE file for details.
@@ -22,7 +22,7 @@ void FeaturesService::begin()
 {
     _server->on(FEATURES_SERVICE_PATH, HTTP_GET, [&](PsychicRequest *request)
                 {
-                    PsychicJsonResponse response = PsychicJsonResponse(request, false, MAX_FEATURES_SIZE);
+                    PsychicJsonResponse response = PsychicJsonResponse(request, false);
                     JsonObject root = response.getRoot();
 
 #if FT_ENABLED(FT_SECURITY)