+++ lib/framework/MqttEndpoint.h
@@ -30,15 +30,13 @@ public:
                  PsychicMqttClient *mqttClient,
                  const String &pubTopic = "",
                  const String &subTopic = "",
-                 bool retain = false,
-                 size_t bufferSize = DEFAULT_BUFFER_SIZE) : _stateReader(stateReader),
-                                                            _stateUpdater(stateUpdater),
-                                                            _statefulService(statefulService),
-                                                            _mqttClient(mqttClient),
-                                                            _pubTopic(pubTopic),
-                                                            _subTopic(subTopic),
-                                                            _retain(retain),
-                                                            _bufferSize(bufferSize)
+                 bool retain = false) : _stateReader(stateReader),
+                                        _stateUpdater(stateUpdater),
+                                        _statefulService(statefulService),
+                                        _mqttClient(mqttClient),
+                                        _pubTopic(pubTopic),
+                                        _subTopic(subTopic),
+                                        _retain(retain)
 
     {
         _statefulService->addUpdateHandler([&](const String &originId)
@@ -95,7 +93,7 @@ public:
         if (_pubTopic.length() > 0 && _mqttClient->connected())
         {
             // serialize to json doc
-            DynamicJsonDocument json(_bufferSize);
+            JsonDocument json;
             JsonObject jsonObject = json.to<JsonObject>();
             _statefulService->read(jsonObject, _stateReader);
 
@@ -116,7 +114,6 @@ public:
 protected:
     StatefulService<T> *_statefulService;
     PsychicMqttClient *_mqttClient;
-    int _bufferSize;
     JsonStateUpdater<T> _stateUpdater;
     JsonStateReader<T> _stateReader;
     String _subTopic;
@@ -136,7 +133,7 @@ protected:
         }
 
         // deserialize from string
-        DynamicJsonDocument json(_bufferSize);
+        JsonDocument json;
         DeserializationError error = deserializeJson(json, payload);
         if (!error && json.is<JsonObject>())
         {